dnsfilexfer/requirements.txt
dnspython>=1.11.1
pycrypto>=2.6.1
